Refactor data structure: consolidate data imports, remove unused files, and reorganize types and utility functions
