+++ src/App.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { CustomAppBar } from "./components/navigation";
 import { HomePage, ExperiencePage, ProjectsPage } from "./components/pages";
 import SocialLinksFooter from "./components/pages/common/SocialLinksFooter";
-import { Pages } from "./data/appData";
+import { Pages } from "./data";
 import { useAppSelector, useAppDispatch } from "./store/hooks";
 import { selectAllDataLoading } from "./store/selectors";
 import { fetchAppData } from "./store/slices/appDataSlice";
